Splash screen and commenting code
