+++ firmware/sserif13.c
@@ -1,0 +1,221 @@
+#include <stdint.h>
+
+// Font size: 8
+// Pixel height: 13
+const uint8_t sserif13_widths[] = {
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+3,3,5,7,6,8,6,2,3,3,4,6,3,3,3,5,6,6,6,6,6,6,6,6,6,6,3,3,6,6,6,6,
+11,7,7,7,8,7,6,8,8,3,5,7,6,9,8,8,7,8,8,7,7,8,7,11,7,7,7,3,5,3,6,6,
+3,6,6,6,6,6,3,6,6,2,2,6,2,8,6,6,6,6,3,5,3,6,6,8,5,5,5,4,2,4,7,3,
+6,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,
+3,3,6,6,6,6,2,6,3,9,4,6,6,3,8,6,4,6,3,3,3,6,6,3,3,3,4,6,8,8,8,6,
+7,7,7,7,7,7,10,7,7,7,7,7,3,3,3,3,8,8,8,8,8,8,8,6,8,8,8,8,8,7,7,6,
+6,6,6,6,6,6,10,6,6,6,6,6,2,4,4,4,6,6,6,6,6,6,6,6,6,6,6,6,6,5,6,5 };
+
+const uint16_t sserif13_offsets[] = {
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,13,26,39,52,65,78,91,104,117,130,143,156,169,182,195,
+208,221,234,247,260,273,286,299,312,325,338,351,364,377,390,403,
+416,442,455,468,481,494,507,520,533,546,559,572,585,598,624,637,
+650,663,676,689,702,715,728,741,767,780,793,806,819,832,845,858,
+871,884,897,910,923,936,949,962,975,988,1001,1014,1027,1040,1053,1066,
+1079,1092,1105,1118,1131,1144,1157,1170,1183,1196,1209,1222,1235,1248,1261,1274,
+1287,1300,1313,1326,1339,1352,1365,1378,1391,1404,1417,1430,1443,1456,1469,1482,
+1495,1508,1521,1534,1547,1560,1573,1586,1599,1612,1625,1638,1651,1664,1677,1690,
+1703,1716,1729,1742,1755,1768,1781,1794,1807,1820,1846,1859,1872,1885,1898,1911,
+1924,1937,1950,1963,1976,1989,2002,2015,2028,2041,2054,2067,2080,2093,2106,2119,
+2132,2145,2158,2171,2184,2197,2210,2236,2249,2262,2275,2288,2301,2314,2327,2340,
+2353,2366,2379,2392,2405,2418,2431,2444,2457,2470,2483,2496,2509,2522,2535,2548,
+2561,2574,2587,2600,2613,2626,2639,2665,2678,2691,2704,2717,2730,2743,2756,2769,
+2782,2795,2808,2821,2834,2847,2860,2873,2886,2899,2912,2925,2938,2951,2964,2977 };
+
+const uint8_t sserif13_data[] = {
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,
+0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x02,0x00,0x00,0x00,0x00,0x12,0x12,0x12,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x24,0x24,0x7e,0x24,0x24,0x24,0x7e,
+0x24,0x24,0x00,0x00,0x00,0x00,0x08,0x1c,0x2a,0x0a,0x0c,0x18,0x28,0x2a,0x1c,0x08,
+0x00,0x00,0x00,0x0c,0x92,0x4c,0x20,0x10,0x08,0x64,0x92,0x60,0x00,0x00,0x00,0x00,
+0x04,0x0a,0x0a,0x04,0x04,0x2a,0x12,0x12,0x2c,0x00,0x00,0x00,0x00,0x02,0x02,0x02,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x02,0x02,0x02,0x02,0x02,
+0x02,0x02,0x02,0x02,0x04,0x00,0x00,0x02,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,
+0x04,0x02,0x00,0x00,0x00,0x0a,0x04,0x0a,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x08,0x08,0x3e,0x08,0x08,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x06,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x02,0x00,0x00,0x00,0x00,0x10,0x10,0x10,0x08,0x08,0x04,0x04,0x02,0x02,0x00,0x00,
+0x00,0x00,0x1c,0x22,0x22,0x22,0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x08,
+0x0e,0x08,0x08,0x08,0x08,0x08,0x08,0x08,0x00,0x00,0x00,0x00,0x1c,0x22,0x20,0x20,
+0x10,0x08,0x04,0x02,0x3e,0x00,0x00,0x00,0x00,0x1c,0x22,0x20,0x20,0x18,0x20,0x20,
+0x22,0x1c,0x00,0x00,0x00,0x00,0x10,0x18,0x18,0x14,0x14,0x12,0x3e,0x10,0x10,0x00,
+0x00,0x00,0x00,0x3e,0x02,0x02,0x1e,0x22,0x20,0x20,0x22,0x1c,0x00,0x00,0x00,0x00,
+0x1c,0x22,0x02,0x02,0x1e,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x3e,0x20,0x10,
+0x10,0x08,0x08,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x1c,0x22,0x22,0x22,0x1c,0x22,
+0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x1c,0x22,0x22,0x22,0x3c,0x20,0x20,0x22,0x1c,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,
+0x10,0x08,0x04,0x02,0x04,0x08,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3e,
+0x00,0x3e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x04,0x08,0x10,0x08,0x04,
+0x02,0x00,0x00,0x00,0x00,0x1c,0x22,0x20,0x20,0x10,0x08,0x08,0x00,0x08,0x00,0x00,
+0x00,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,0x04,0x02,0xe2,0x04,0x92,0x04,0x92,0x04,
+0x62,0x07,0x04,0x00,0x0c,0x00,0xf0,0x01,0x00,0x00,0x00,0x00,0x08,0x08,0x14,0x14,
+0x22,0x22,0x3e,0x41,0x41,0x00,0x00,0x00,0x00,0x1e,0x22,0x22,0x22,0x1e,0x22,0x22,
+0x22,0x1e,0x00,0x00,0x00,0x00,0x3c,0x42,0x02,0x02,0x02,0x02,0x02,0x42,0x3c,0x00,
+0x00,0x00,0x00,0x1e,0x22,0x42,0x42,0x42,0x42,0x42,0x22,0x1e,0x00,0x00,0x00,0x00,
+0x3e,0x02,0x02,0x02,0x1e,0x02,0x02,0x02,0x3e,0x00,0x00,0x00,0x00,0x3e,0x02,0x02,
+0x02,0x1e,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x3c,0x42,0x02,0x02,0x72,0x42,
+0x42,0x62,0x5c,0x00,0x00,0x00,0x00,0x42,0x42,0x42,0x42,0x7e,0x42,0x42,0x42,0x42,
+0x00,0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
+0x00,0x08,0x08,0x08,0x08,0x08,0x08,0x09,0x09,0x06,0x00,0x00,0x00,0x00,0x22,0x12,
+0x0a,0x06,0x06,0x0a,0x12,0x22,0x42,0x00,0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,
+0x02,0x02,0x02,0x3e,0x00,0x00,0x00,0x00,0x00,0x00,0x82,0x00,0x82,0x00,0xc6,0x00,
+0xc6,0x00,0xaa,0x00,0xaa,0x00,0x92,0x00,0x92,0x00,0x82,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x42,0x46,0x46,0x4a,0x4a,0x52,0x62,0x62,0x42,0x00,0x00,0x00,0x00,0x3c,
+0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,0x00,0x3e,0x42,0x42,0x42,
+0x3e,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x52,
+0x62,0x3c,0x40,0x00,0x00,0x00,0x3e,0x42,0x42,0x42,0x3e,0x42,0x42,0x42,0x42,0x00,
+0x00,0x00,0x00,0x1c,0x22,0x02,0x02,0x1c,0x20,0x20,0x22,0x1c,0x00,0x00,0x00,0x00,
+0x3e,0x08,0x08,0x08,0x08,0x08,0x08,0x08,0x08,0x00,0x00,0x00,0x00,0x42,0x42,0x42,
+0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,0x00,0x41,0x41,0x22,0x22,0x22,0x14,
+0x14,0x08,0x08,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x04,0x01,0x04,0x22,0x02,0x22,
+0x02,0x22,0x02,0x54,0x01,0x54,0x01,0x88,0x00,0x88,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x41,0x41,0x22,0x14,0x08,0x14,0x22,0x41,0x41,0x00,0x00,0x00,0x00,0x41,0x41,
+0x22,0x14,0x08,0x08,0x08,0x08,0x08,0x00,0x00,0x00,0x00,0x7f,0x40,0x20,0x10,0x08,
+0x04,0x02,0x01,0x7f,0x00,0x00,0x00,0x00,0x06,0x02,0x02,0x02,0x02,0x02,0x02,0x02,
+0x02,0x02,0x06,0x00,0x00,0x02,0x02,0x02,0x04,0x04,0x08,0x08,0x10,0x10,0x00,0x00,
+0x00,0x00,0x06,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x06,0x00,0x08,0x14,
+0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x3f,0x00,0x00,0x02,0x04,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1c,0x20,0x3c,0x22,0x22,0x3c,0x00,
+0x00,0x00,0x00,0x02,0x02,0x02,0x1e,0x22,0x22,0x22,0x22,0x1e,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x1c,0x22,0x02,0x02,0x22,0x1c,0x00,0x00,0x00,0x00,0x20,0x20,0x20,
+0x3c,0x22,0x22,0x22,0x22,0x3c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1c,0x22,0x3e,
+0x02,0x22,0x1c,0x00,0x00,0x00,0x00,0x04,0x02,0x02,0x06,0x02,0x02,0x02,0x02,0x02,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x22,0x22,0x22,0x22,0x3c,0x20,0x1e,0x00,
+0x00,0x02,0x02,0x02,0x1a,0x26,0x22,0x22,0x22,0x22,0x00,0x00,0x00,0x00,0x02,0x00,
+0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x02,
+0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x02,0x02,0x02,0x12,0x0a,0x06,0x0a,0x12,
+0x22,0x00,0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x6e,0x92,0x92,0x92,0x92,0x92,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1a,0x26,0x22,0x22,0x22,0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1c,
+0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1e,0x22,0x22,0x22,
+0x22,0x1e,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x3c,0x22,0x22,0x22,0x22,0x3c,0x20,
+0x20,0x00,0x00,0x00,0x00,0x00,0x06,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x0c,0x12,0x04,0x08,0x12,0x0c,0x00,0x00,0x00,0x00,0x00,0x02,0x02,
+0x06,0x02,0x02,0x02,0x02,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x22,0x22,0x22,
+0x22,0x32,0x2c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x22,0x22,0x14,0x14,0x08,0x08,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x92,0x92,0xaa,0xaa,0x44,0x44,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x12,0x12,0x0c,0x0c,0x12,0x12,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x12,0x12,0x12,0x12,0x0c,0x04,0x04,0x03,0x00,0x00,0x00,0x00,0x00,0x1e,0x10,
+0x08,0x04,0x02,0x1e,0x00,0x00,0x00,0x08,0x04,0x04,0x04,0x04,0x02,0x04,0x04,0x04,
+0x04,0x08,0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,
+0x00,0x02,0x04,0x04,0x04,0x04,0x08,0x04,0x04,0x04,0x04,0x02,0x00,0x00,0x00,0x00,
+0x4c,0x32,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x1c,0x22,0x22,0x0f,0x02,0x07,0x22,
+0x22,0x1c,0x00,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,
+0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
+0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,
+0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,
+0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,
+0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,
+0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,
+0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,
+0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,
+0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,
+0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,
+0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,
+0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,
+0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x06,0x06,0x06,
+0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,
+0x02,0x00,0x00,0x00,0x00,0x10,0x1c,0x22,0x02,0x02,0x22,0x1c,0x04,0x00,0x00,0x00,
+0x0c,0x12,0x02,0x04,0x0e,0x04,0x02,0x02,0x3e,0x00,0x00,0x00,0x00,0x22,0x1c,0x22,
+0x22,0x22,0x1c,0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x22,0x22,0x22,0x14,0x3e,0x08,
+0x3e,0x08,0x08,0x00,0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x00,0x02,0x02,0x02,
+0x02,0x02,0x00,0x00,0x1c,0x22,0x04,0x0a,0x12,0x24,0x28,0x10,0x22,0x1c,0x00,0x00,
+0x00,0x00,0x05,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x78,0x00,0x84,0x00,0x32,0x01,0x0a,0x01,0x32,0x01,0x84,0x00,0x78,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x0c,0x0a,0x0e,0x00,0x0e,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x28,0x14,0x0a,0x14,0x28,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3e,0x20,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x44,0xba,0x9a,
+0xaa,0x44,0x38,0x00,0x00,0x00,0x00,0x00,0x3f,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x0a,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x08,0x08,0x3e,0x08,0x08,0x00,0x3e,0x00,0x00,0x00,0x00,
+0x03,0x04,0x02,0x01,0x07,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x03,0x04,0x02,
+0x04,0x03,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x12,0x12,0x12,0x12,0x12,0x2e,
+0x02,0x02,0x00,0x00,0x3c,0x3e,0x3e,0x3e,0x3c,0x28,0x28,0x28,0x28,0x28,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x06,0x00,0x00,0x04,0x06,0x04,0x04,0x04,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0a,0x0a,0x0e,0x00,0x0e,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0a,0x14,0x28,0x14,0x0a,0x00,0x00,
+0x00,0x00,0x04,0x86,0x44,0x24,0x94,0xc8,0xa4,0xe2,0x80,0x00,0x00,0x00,0x00,0x04,
+0x86,0x44,0x24,0x74,0x88,0x44,0x22,0xe0,0x00,0x00,0x00,0x00,0x06,0x88,0x44,0x28,
+0x96,0xc8,0xa4,0xe2,0x80,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,0x08,0x08,0x04,
+0x02,0x02,0x22,0x1c,0x04,0x08,0x08,0x08,0x14,0x14,0x22,0x22,0x3e,0x41,0x41,0x00,
+0x00,0x10,0x08,0x08,0x08,0x14,0x14,0x22,0x22,0x3e,0x41,0x41,0x00,0x00,0x08,0x14,
+0x00,0x08,0x14,0x14,0x22,0x22,0x3e,0x41,0x41,0x00,0x00,0x2c,0x1a,0x08,0x08,0x14,
+0x14,0x22,0x22,0x3e,0x41,0x41,0x00,0x00,0x22,0x00,0x08,0x08,0x14,0x14,0x22,0x22,
+0x3e,0x41,0x41,0x00,0x00,0x08,0x14,0x08,0x08,0x14,0x14,0x22,0x22,0x3e,0x41,0x41,
+0x00,0x00,0x00,0x00,0x00,0x00,0xf8,0x01,0x18,0x00,0x14,0x00,0x14,0x00,0xf2,0x00,
+0x1e,0x00,0x12,0x00,0x11,0x00,0xf1,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x42,
+0x02,0x02,0x02,0x02,0x02,0x42,0x3c,0x10,0x18,0x04,0x08,0x3e,0x02,0x02,0x02,0x1e,
+0x02,0x02,0x02,0x3e,0x00,0x00,0x10,0x08,0x3e,0x02,0x02,0x02,0x1e,0x02,0x02,0x02,
+0x3e,0x00,0x00,0x08,0x14,0x3e,0x02,0x02,0x02,0x1e,0x02,0x02,0x02,0x3e,0x00,0x00,
+0x22,0x00,0x3e,0x02,0x02,0x02,0x1e,0x02,0x02,0x02,0x3e,0x00,0x00,0x01,0x02,0x00,
+0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x04,0x02,0x00,0x02,0x02,0x02,
+0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x02,0x05,0x00,0x02,0x02,0x02,0x02,0x02,0x02,
+0x02,0x02,0x00,0x00,0x05,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,
+0x00,0x00,0x00,0x1e,0x22,0x42,0x42,0x4f,0x42,0x42,0x22,0x1e,0x00,0x00,0x4c,0x32,
+0x42,0x46,0x46,0x4a,0x4a,0x52,0x62,0x62,0x42,0x00,0x00,0x08,0x10,0x3c,0x42,0x42,
+0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x10,0x08,0x3c,0x42,0x42,0x42,0x42,0x42,
+0x42,0x42,0x3c,0x00,0x00,0x18,0x24,0x3c,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,
+0x00,0x00,0x4c,0x32,0x3c,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x24,
+0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x22,0x14,0x08,0x14,0x22,0x00,0x00,0x00,0x00,0x00,0x7c,0x62,0x52,0x52,0x4a,
+0x4a,0x46,0x46,0x3e,0x00,0x00,0x08,0x10,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x42,
+0x3c,0x00,0x00,0x10,0x08,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,
+0x18,0x24,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x24,0x00,0x42,
+0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x10,0x08,0x41,0x41,0x22,0x14,
+0x08,0x08,0x08,0x08,0x08,0x00,0x00,0x00,0x00,0x02,0x02,0x3e,0x42,0x42,0x42,0x3e,
+0x02,0x02,0x00,0x00,0x00,0x00,0x0c,0x12,0x12,0x12,0x1a,0x22,0x22,0x22,0x1a,0x00,
+0x00,0x00,0x00,0x04,0x08,0x00,0x1c,0x20,0x3c,0x22,0x22,0x3c,0x00,0x00,0x00,0x00,
+0x10,0x08,0x00,0x1c,0x20,0x3c,0x22,0x22,0x3c,0x00,0x00,0x00,0x00,0x08,0x14,0x00,
+0x1c,0x20,0x3c,0x22,0x22,0x3c,0x00,0x00,0x00,0x00,0x2c,0x1a,0x00,0x1c,0x20,0x3c,
+0x22,0x22,0x3c,0x00,0x00,0x00,0x00,0x24,0x00,0x00,0x1c,0x20,0x3c,0x22,0x22,0x3c,
+0x00,0x00,0x00,0x08,0x14,0x08,0x00,0x1c,0x20,0x3c,0x22,0x22,0x3c,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xdc,0x01,0x20,0x02,0xfc,0x03,0x22,
+0x00,0x22,0x02,0xdc,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1c,0x22,
+0x02,0x02,0x22,0x1c,0x08,0x0c,0x00,0x00,0x04,0x08,0x00,0x1c,0x22,0x3e,0x02,0x22,
+0x1c,0x00,0x00,0x00,0x00,0x10,0x08,0x00,0x1c,0x22,0x3e,0x02,0x22,0x1c,0x00,0x00,
+0x00,0x00,0x08,0x14,0x00,0x1c,0x22,0x3e,0x02,0x22,0x1c,0x00,0x00,0x00,0x00,0x22,
+0x00,0x00,0x1c,0x22,0x3e,0x02,0x22,0x1c,0x00,0x00,0x00,0x00,0x01,0x02,0x00,0x02,
+0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x08,0x04,0x00,0x04,0x04,0x04,0x04,
+0x04,0x04,0x00,0x00,0x00,0x00,0x04,0x0a,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x00,
+0x00,0x00,0x00,0x0a,0x00,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,
+0x14,0x08,0x14,0x10,0x3c,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x2c,0x1a,0x00,
+0x1a,0x26,0x22,0x22,0x22,0x22,0x00,0x00,0x00,0x00,0x04,0x08,0x00,0x1c,0x22,0x22,
+0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x10,0x08,0x00,0x1c,0x22,0x22,0x22,0x22,0x1c,
+0x00,0x00,0x00,0x00,0x08,0x14,0x00,0x1c,0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,
+0x00,0x2c,0x1a,0x00,0x1c,0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x22,0x00,
+0x00,0x1c,0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,
+0x3e,0x00,0x08,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x32,0x2a,0x2a,0x26,
+0x1e,0x00,0x00,0x00,0x00,0x04,0x08,0x00,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,
+0x00,0x00,0x10,0x08,0x00,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,0x00,0x08,
+0x14,0x00,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,0x00,0x22,0x00,0x00,0x22,
+0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,0x00,0x08,0x04,0x00,0x12,0x12,0x12,0x12,
+0x0c,0x04,0x04,0x03,0x00,0x00,0x02,0x02,0x02,0x1e,0x22,0x22,0x22,0x22,0x1e,0x02,
+0x02,0x00,0x00,0x12,0x00,0x00,0x12,0x12,0x12,0x12,0x0c,0x04,0x04,0x03 };
+const font_def_t sserif13 = { 256, 13, sserif13_widths, sserif13_offsets, sserif13_data };
