+++ firmware/sserif16.h
@@ -1,311 +1,10 @@
-// Font size: 10
-// Pixel height: 16
-const uint8_t sserif16_widths[] = {
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-3,3,5,7,7,12,9,3,4,4,5,7,3,4,3,4,7,7,7,7,7,7,7,7,7,7,3,3,7,7,7,7,
-14,9,9,9,10,9,8,10,10,3,7,8,7,11,10,10,9,10,10,9,9,10,9,13,8,9,8,4,4,4,7,7,
-4,8,8,7,8,8,3,8,7,3,3,7,3,11,7,8,8,8,4,7,3,7,7,9,6,7,6,5,3,5,9,4,
-7,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,3,3,4,4,4,4,4,4,4,4,4,4,4,4,4,
-3,3,7,7,7,7,3,7,4,8,5,7,7,4,8,7,4,7,4,4,4,7,7,3,4,4,5,7,9,9,9,7,
-9,9,9,9,9,9,13,9,9,9,9,9,3,3,3,3,9,10,10,10,10,10,10,7,10,10,10,10,10,9,9,8,
-8,8,8,8,8,8,13,7,8,8,8,8,3,3,3,3,8,7,8,8,8,8,8,7,8,7,7,7,7,7,8,7 };
+#ifndef SSERIF16_H
+#define SSERIF16_H
 
-const uint16_t sserif16_offsets[] = {
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,16,32,48,64,80,112,144,160,176,192,208,224,240,256,272,
-288,304,320,336,352,368,384,400,416,432,448,464,480,496,512,528,
-544,576,608,640,672,704,736,752,784,816,832,848,864,880,912,944,
-976,1008,1040,1072,1104,1136,1168,1200,1232,1248,1280,1296,1312,1328,1344,1360,
-1376,1392,1408,1424,1440,1456,1472,1488,1504,1520,1536,1552,1568,1584,1616,1632,
-1648,1664,1680,1696,1712,1728,1744,1760,1792,1808,1824,1840,1856,1872,1888,1920,
-1936,1952,1968,1984,2000,2016,2032,2048,2064,2080,2096,2112,2128,2144,2160,2176,
-2192,2208,2224,2240,2256,2272,2288,2304,2320,2336,2352,2368,2384,2400,2416,2432,
-2448,2464,2480,2496,2512,2528,2544,2560,2576,2592,2608,2624,2640,2656,2672,2688,
-2704,2720,2736,2752,2768,2784,2800,2816,2832,2848,2864,2880,2896,2928,2960,2992,
-3008,3040,3072,3104,3136,3168,3200,3232,3264,3296,3328,3360,3392,3408,3424,3440,
-3456,3488,3520,3552,3584,3616,3648,3680,3696,3728,3760,3792,3824,3856,3888,3920,
-3936,3952,3968,3984,4000,4016,4032,4064,4080,4096,4112,4128,4144,4160,4176,4192,
-4208,4224,4240,4256,4272,4288,4304,4320,4336,4352,4368,4384,4400,4416,4432,4448 };
+extern const uint8_t sserif16_widths[];
+extern const uint16_t sserif16_offseets[];
+extern const uint8_t sserif16_data[];
+extern const font_def_t sserif16;
 
-const uint8_t sserif16_data[] = {
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x0a,0x0a,0x0a,0x0a,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x28,0x28,0x28,0x7e,0x14,0x14,0x3f,0x0a,0x0a,0x0a,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x1c,0x2a,0x2a,0x0a,0x1c,0x28,0x28,0x2a,0x1c,0x08,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x0c,0x00,0x12,0x02,0x12,0x01,0x8c,0x00,0x40,0x00,
-0x20,0x00,0x10,0x03,0x88,0x04,0x84,0x04,0x00,0x03,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x18,0x00,0x24,0x00,0x24,0x00,0x24,0x00,0x18,0x00,
-0x14,0x00,0x22,0x00,0xa2,0x00,0x42,0x00,0xbc,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x04,0x04,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x04,0x04,0x08,
-0x00,0x00,0x00,0x01,0x02,0x02,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x02,0x02,0x01,
-0x00,0x00,0x00,0x04,0x15,0x0e,0x1b,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x08,0x3e,0x08,0x08,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x01,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0f,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x08,0x08,0x04,0x04,0x04,0x02,0x02,0x02,0x01,0x01,0x01,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x22,0x22,0x22,0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x0e,0x08,0x08,0x08,0x08,0x08,0x08,0x08,0x08,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x22,0x20,0x10,0x08,0x04,0x02,0x02,0x3e,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x20,0x20,0x18,0x20,0x20,0x20,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x20,0x30,0x28,0x28,0x24,0x22,0x3e,0x20,0x20,0x20,0x00,0x00,0x00,
-0x00,0x00,0x00,0x3e,0x02,0x02,0x02,0x1e,0x22,0x20,0x20,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x02,0x02,0x1a,0x26,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x3e,0x20,0x20,0x10,0x10,0x08,0x08,0x04,0x04,0x04,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x22,0x22,0x1c,0x22,0x22,0x22,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x22,0x22,0x32,0x2c,0x20,0x20,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x02,0x01,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x30,0x18,0x0c,0x06,0x0c,0x18,0x30,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3e,0x00,0x3e,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x06,0x0c,0x18,0x30,0x18,0x0c,0x06,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x22,0x20,0x10,0x08,0x08,0x08,0x00,0x08,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xe0,0x01,0x18,0x06,0x04,0x08,0x04,0x08,0xc2,0x13,0x22,0x12,
-0x22,0x12,0x22,0x11,0xc4,0x0e,0x04,0x00,0x18,0x0c,0xe0,0x03,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x7e,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x7e,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x7e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x00,0x02,0x00,
-0x02,0x00,0x02,0x00,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x7e,0x00,0x82,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x82,0x00,0x7e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0xfe,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x7e,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0xfe,0x02,0x02,0x02,0x7e,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x00,0x02,0x00,
-0xc2,0x01,0x02,0x01,0x02,0x01,0x84,0x01,0x78,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0xfe,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x20,0x20,0x20,0x20,0x20,0x20,0x20,0x22,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x42,0x22,0x12,0x0a,0x06,0x0a,0x12,0x22,0x42,0x82,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x7e,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x06,0x03,0x06,0x03,0x8a,0x02,
-0x8a,0x02,0x52,0x02,0x52,0x02,0x22,0x02,0x22,0x02,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x01,0x06,0x01,0x0a,0x01,0x0a,0x01,0x12,0x01,
-0x22,0x01,0x42,0x01,0x42,0x01,0x82,0x01,0x02,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x7e,0x00,0x82,0x00,0x02,0x01,0x02,0x01,0x82,0x00,
-0x7e,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x42,0x01,0x84,0x00,0x78,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x7e,0x00,0x82,0x00,0x02,0x01,0x02,0x01,0x82,0x00,
-0xfe,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x7c,0x00,0x82,0x00,0x82,0x00,0x02,0x00,0x1c,0x00,
-0x60,0x00,0x80,0x00,0x82,0x00,0x82,0x00,0x7c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0xff,0x01,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,
-0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x01,0x01,0x01,0x82,0x00,0x82,0x00,0x44,0x00,
-0x44,0x00,0x28,0x00,0x28,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x41,0x10,0x41,0x10,0x41,0x10,0xa2,0x08,0xa2,0x08,
-0x14,0x05,0x14,0x05,0x14,0x05,0x08,0x02,0x08,0x02,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x81,0x81,0x42,0x24,0x18,0x18,0x24,0x42,0x81,0x81,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x01,0x01,0x01,0x82,0x00,0x44,0x00,0x28,0x00,
-0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x7e,0x40,0x20,0x20,0x10,0x08,0x04,0x04,0x02,0x7e,0x00,0x00,0x00,
-0x00,0x00,0x00,0x0e,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x0e,
-0x00,0x00,0x00,0x01,0x01,0x01,0x02,0x02,0x02,0x04,0x04,0x04,0x08,0x08,0x08,0x00,
-0x00,0x00,0x00,0x07,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x07,
-0x00,0x00,0x08,0x14,0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7f,0x00,
-0x00,0x00,0x00,0x02,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x3a,0x46,0x42,0x42,0x42,0x46,0x3a,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x42,0x02,0x02,0x02,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x40,0x40,0x40,0x5c,0x62,0x42,0x42,0x42,0x62,0x5c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x42,0x42,0x7e,0x02,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x02,0x02,0x07,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x5c,0x62,0x42,0x42,0x42,0x62,0x5c,0x40,0x40,0x3c,
-0x00,0x00,0x00,0x02,0x02,0x02,0x1a,0x26,0x22,0x22,0x22,0x22,0x22,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x01,
-0x00,0x00,0x00,0x02,0x02,0x02,0x22,0x12,0x0a,0x0e,0x12,0x22,0x42,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x9a,0x01,0x66,0x02,
-0x22,0x02,0x22,0x02,0x22,0x02,0x22,0x02,0x22,0x02,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x1a,0x26,0x22,0x22,0x22,0x22,0x22,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3a,0x46,0x42,0x42,0x42,0x46,0x3a,0x02,0x02,0x02,
-0x00,0x00,0x00,0x00,0x00,0x00,0x5c,0x62,0x42,0x42,0x42,0x62,0x5c,0x40,0x40,0x40,
-0x00,0x00,0x00,0x00,0x00,0x00,0x0a,0x06,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x1c,0x22,0x02,0x1c,0x20,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x02,0x02,0x07,0x02,0x02,0x02,0x02,0x02,0x04,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x22,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x11,0x01,0x11,0x01,
-0xaa,0x00,0xaa,0x00,0xaa,0x00,0x44,0x00,0x44,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x21,0x12,0x0c,0x0c,0x0c,0x12,0x21,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x08,0x04,0x03,
-0x00,0x00,0x00,0x00,0x00,0x00,0x1e,0x10,0x08,0x04,0x04,0x02,0x1e,0x00,0x00,0x00,
-0x00,0x00,0x00,0x18,0x04,0x04,0x04,0x04,0x04,0x03,0x04,0x04,0x04,0x04,0x04,0x18,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,
-0x00,0x00,0x00,0x03,0x04,0x04,0x04,0x04,0x04,0x18,0x04,0x04,0x04,0x04,0x04,0x03,
-0x00,0x00,0x00,0x00,0x00,0x00,0x8c,0x00,0x92,0x00,0x62,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x1c,0x22,0x22,0x0f,0x02,0x07,0x02,0x22,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x06,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,
-0x00,0x00,0x00,0x00,0x00,0x20,0x1c,0x32,0x12,0x0a,0x0a,0x26,0x1c,0x02,0x00,0x00,
-0x00,0x00,0x00,0x0c,0x12,0x02,0x02,0x04,0x1e,0x04,0x04,0x02,0x3e,0x00,0x00,0x00,
-0x00,0x00,0x00,0x22,0x1c,0x14,0x1c,0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x41,0x41,0x22,0x14,0x7f,0x08,0x7f,0x08,0x08,0x08,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,
-0x00,0x00,0x00,0x1c,0x22,0x04,0x0c,0x12,0x24,0x18,0x10,0x22,0x1c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x09,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x3c,0x42,0xb9,0x85,0x85,0xb9,0x42,0x3c,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x08,0x0c,0x0a,0x0c,0x00,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x28,0x14,0x0a,0x14,0x28,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3e,0x20,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0f,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x3c,0x42,0x9d,0xa5,0x9d,0xa5,0x42,0x3c,0x00,0x00,0x00,0x00,0x00,
-0x00,0x7f,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x0a,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x08,0x3e,0x08,0x08,0x00,0x3e,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x09,0x04,0x02,0x0f,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x06,0x09,0x04,0x09,0x06,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x22,0x22,0x22,0x22,0x22,0x22,0x22,0x5e,0x02,0x02,0x02,
-0x00,0x00,0x00,0x3c,0x2e,0x2e,0x2e,0x2e,0x2c,0x28,0x28,0x28,0x28,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x08,0x06,
-0x00,0x00,0x00,0x04,0x06,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x0a,0x0a,0x0a,0x04,0x00,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0a,0x14,0x28,0x14,0x0a,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x42,0x00,0x43,0x00,0x22,0x00,0x22,0x00,0x12,0x00,
-0x10,0x01,0x88,0x01,0x48,0x01,0xe4,0x01,0x04,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x42,0x00,0x43,0x00,0x22,0x00,0x22,0x00,0x12,0x00,
-0xd0,0x00,0x28,0x01,0x88,0x00,0x44,0x00,0xe4,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x46,0x00,0x49,0x00,0x24,0x00,0x29,0x00,0x16,0x00,
-0x10,0x01,0x88,0x01,0x48,0x01,0xe4,0x01,0x04,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,0x08,0x08,0x08,0x04,0x02,0x22,0x22,0x1c,
-0x08,0x00,0x10,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x20,0x00,0x10,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x10,0x00,0x28,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x9c,0x00,0x72,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x44,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x10,0x00,0x28,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x28,0x00,0x28,0x00,0x44,0x00,
-0x44,0x00,0xfe,0x00,0x82,0x00,0x01,0x01,0x01,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0xf0,0x0f,0x30,0x00,0x28,0x00,0x28,0x00,0xe4,0x07,
-0x24,0x00,0x3e,0x00,0x22,0x00,0x21,0x00,0xe1,0x0f,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x00,0x02,0x00,
-0x02,0x00,0x02,0x00,0x02,0x01,0x84,0x00,0x78,0x00,0x20,0x00,0x40,0x00,0x30,0x00,
-0x08,0x00,0x10,0x00,0x00,0x00,0xfe,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x7e,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x20,0x00,0x10,0x00,0x00,0x00,0xfe,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x7e,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x10,0x00,0x28,0x00,0x00,0x00,0xfe,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x7e,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x44,0x00,0x00,0x00,0x00,0x00,0xfe,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x7e,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x01,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x04,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x02,0x05,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x05,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3e,0x00,0x42,0x00,0x82,0x00,0x82,0x00,0x8f,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x42,0x00,0x3e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x98,0x00,0x64,0x00,0x00,0x00,0x02,0x01,0x06,0x01,0x0a,0x01,0x0a,0x01,0x12,0x01,
-0x22,0x01,0x42,0x01,0x42,0x01,0x82,0x01,0x02,0x01,0x00,0x00,0x00,0x00,0x00,0x00,
-0x08,0x00,0x10,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x40,0x00,0x20,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x30,0x00,0x48,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x98,0x00,0x64,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x84,0x00,0x00,0x00,0x00,0x00,0x78,0x00,0x84,0x00,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x22,0x14,0x08,0x14,0x22,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x78,0x01,0x84,0x00,0x42,0x01,0x42,0x01,0x22,0x01,
-0x12,0x01,0x0a,0x01,0x0a,0x01,0x84,0x00,0x7a,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x10,0x00,0x20,0x00,0x00,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x20,0x00,0x10,0x00,0x00,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x30,0x00,0x48,0x00,0x48,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x84,0x00,0x00,0x00,0x00,0x00,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,0x02,0x01,
-0x02,0x01,0x02,0x01,0x02,0x01,0x84,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x20,0x00,0x10,0x00,0x00,0x00,0x01,0x01,0x01,0x01,0x82,0x00,0x44,0x00,0x28,0x00,
-0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x02,0x00,0x3e,0x00,0x42,0x00,0x82,0x00,
-0x82,0x00,0x42,0x00,0x3e,0x00,0x02,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x3c,0x42,0x42,0x42,0x3a,0x42,0x42,0x42,0x42,0x3a,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x10,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x00,0x10,0x08,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x00,0x18,0x24,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x00,0x4c,0x32,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x00,0x24,0x00,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x18,0x24,0x18,0x00,0x3c,0x42,0x7c,0x42,0x42,0x42,0xbc,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xbc,0x07,0x42,0x08,
-0x7c,0x08,0xc2,0x0f,0x42,0x00,0x42,0x08,0xbc,0x07,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3c,0x42,0x02,0x02,0x02,0x42,0x3c,0x10,0x20,0x18,
-0x00,0x00,0x00,0x08,0x10,0x00,0x3c,0x42,0x42,0x7e,0x02,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x10,0x08,0x00,0x3c,0x42,0x42,0x7e,0x02,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x18,0x24,0x00,0x3c,0x42,0x42,0x7e,0x02,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x24,0x00,0x00,0x3c,0x42,0x42,0x7e,0x02,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x01,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x02,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x02,0x05,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x05,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,
-0x00,0x00,0x00,0x28,0x10,0x28,0x20,0x7c,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x2c,0x1a,0x00,0x1a,0x26,0x22,0x22,0x22,0x22,0x22,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x10,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x10,0x08,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x18,0x24,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x4c,0x32,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x24,0x00,0x00,0x3c,0x42,0x42,0x42,0x42,0x42,0x3c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,0x3e,0x00,0x08,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x40,0x3c,0x62,0x52,0x52,0x4a,0x46,0x3c,0x02,0x00,0x00,
-0x00,0x00,0x00,0x04,0x08,0x00,0x22,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x10,0x08,0x00,0x22,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x14,0x00,0x22,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x22,0x00,0x00,0x22,0x22,0x22,0x22,0x22,0x32,0x2c,0x00,0x00,0x00,
-0x00,0x00,0x00,0x10,0x08,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x08,0x04,0x03,
-0x00,0x00,0x00,0x02,0x02,0x02,0x3a,0x46,0x42,0x42,0x42,0x46,0x3a,0x02,0x02,0x02,
-0x00,0x00,0x00,0x22,0x00,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x08,0x04,0x03 };
-const font_def_t sserif16 = { 256, 16, sserif16_widths, sserif16_offsets, sserif16_data };
+
+#endif
