+++ firmware/sserif20.h
@@ -1,474 +1,9 @@
-// Font size: 12
-// Pixel height: 20
-const uint8_t sserif20_widths[] = {
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-4,4,6,9,9,14,11,3,5,5,6,9,4,5,4,4,9,9,9,9,9,9,9,9,9,9,4,4,9,9,9,9,
-16,11,11,11,12,11,10,13,12,5,8,10,9,13,11,12,10,12,12,11,9,12,11,15,11,11,10,4,4,4,7,9,
-5,9,9,8,9,9,5,9,9,3,3,8,3,13,9,9,9,9,5,8,5,9,7,11,7,7,8,5,5,5,9,5,
-9,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,4,4,5,5,5,5,5,5,5,5,5,5,5,5,5,
-4,4,9,9,9,9,5,9,5,11,6,9,9,5,11,9,5,9,5,5,5,8,8,4,5,5,6,9,12,12,12,9,
-11,11,11,11,11,11,16,11,11,11,11,11,5,5,5,5,12,11,12,12,12,12,12,9,12,12,12,12,12,11,10,10,
-9,9,9,9,9,9,15,8,9,9,9,9,3,3,3,3,9,9,9,9,9,9,9,9,9,9,9,9,9,7,9,7 };
+#ifndef SSERIF20_H
+#define SSERIF20_H
 
-const uint16_t sserif20_offsets[] = {
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,20,40,60,100,140,180,220,240,260,280,300,340,360,380,400,
-420,460,500,540,580,620,660,700,740,780,820,840,860,900,940,980,
-1020,1060,1100,1140,1180,1220,1260,1300,1340,1380,1400,1420,1460,1500,1540,1580,
-1620,1660,1700,1740,1780,1820,1860,1900,1940,1980,2020,2060,2080,2100,2120,2140,
-2180,2200,2240,2280,2300,2340,2380,2400,2440,2480,2500,2520,2540,2560,2600,2640,
-2680,2720,2760,2780,2800,2820,2860,2880,2920,2940,2960,2980,3000,3020,3040,3080,
-3100,3140,3160,3180,3200,3220,3240,3260,3280,3300,3320,3340,3360,3380,3400,3420,
-3440,3460,3480,3500,3520,3540,3560,3580,3600,3620,3640,3660,3680,3700,3720,3740,
-3760,3780,3800,3840,3880,3920,3960,3980,4020,4040,4080,4100,4140,4180,4200,4240,
-4280,4300,4340,4360,4380,4400,4420,4440,4460,4480,4500,4520,4560,4600,4640,4680,
-4720,4760,4800,4840,4880,4920,4960,5000,5040,5080,5120,5160,5200,5220,5240,5260,
-5280,5320,5360,5400,5440,5480,5520,5560,5600,5640,5680,5720,5760,5800,5840,5880,
-5920,5960,6000,6040,6080,6120,6160,6200,6220,6260,6300,6340,6380,6400,6420,6440,
-6460,6500,6540,6580,6620,6660,6700,6740,6780,6820,6860,6900,6940,6980,7000,7040 };
+extern const uint8_t sserif20_widths[];
+extern const uint16_t sserif20_offsets[];
+extern const uint8_t sserif20_data[];
+extern const font_def_t sserif20;
 
-const uint8_t sserif20_data[] = {
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,
-0x04,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x12,0x12,0x12,0x12,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x48,0x00,0x48,0x00,0x48,0x00,0xfe,0x00,0x24,0x00,0x24,0x00,
-0x24,0x00,0x24,0x00,0x7f,0x00,0x12,0x00,0x12,0x00,0x12,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x38,0x00,0x54,0x00,
-0x92,0x00,0x12,0x00,0x14,0x00,0x38,0x00,0x50,0x00,0x90,0x00,0x90,0x00,0x92,0x00,
-0x54,0x00,0x38,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x1c,0x10,0x22,0x08,0x22,0x04,0x22,0x02,0x1c,0x01,0x80,0x00,
-0x40,0x00,0x20,0x0e,0x10,0x11,0x08,0x11,0x04,0x11,0x02,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x18,0x00,0x24,0x00,
-0x24,0x00,0x24,0x00,0x18,0x00,0x18,0x00,0x24,0x00,0x42,0x02,0x82,0x02,0x02,0x01,
-0x84,0x02,0x78,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x08,0x04,0x04,0x04,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x04,
-0x04,0x04,0x08,0x00,0x00,0x00,0x00,0x02,0x04,0x04,0x04,0x08,0x08,0x08,0x08,0x08,
-0x08,0x08,0x08,0x04,0x04,0x04,0x02,0x00,0x00,0x00,0x00,0x08,0x2a,0x1c,0x14,0x22,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x10,0x00,
-0xfe,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x1f,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x08,0x08,0x08,0x08,0x04,0x04,0x04,0x04,0x02,0x02,0x02,0x02,
-0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,
-0x44,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x44,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x10,0x00,0x1c,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,
-0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,
-0x82,0x00,0x82,0x00,0x80,0x00,0x40,0x00,0x20,0x00,0x10,0x00,0x08,0x00,0x04,0x00,
-0x02,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0x40,0x00,0x38,0x00,
-0x40,0x00,0x80,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x60,0x00,
-0x60,0x00,0x50,0x00,0x48,0x00,0x48,0x00,0x44,0x00,0x42,0x00,0xfe,0x00,0x40,0x00,
-0x40,0x00,0x40,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xfe,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x3a,0x00,0x46,0x00,
-0x82,0x00,0x80,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,
-0x82,0x00,0x02,0x00,0x3a,0x00,0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xfe,0x00,0x80,0x00,0x80,0x00,0x40,0x00,0x40,0x00,0x20,0x00,
-0x20,0x00,0x20,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,
-0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0xc4,0x00,0xb8,0x00,0x80,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xc0,0x00,0x30,0x00,0x0c,0x00,
-0x02,0x00,0x0c,0x00,0x30,0x00,0xc0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x00,0x18,0x00,0x60,0x00,
-0x80,0x00,0x60,0x00,0x18,0x00,0x06,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,
-0x82,0x00,0x82,0x00,0x80,0x00,0x40,0x00,0x20,0x00,0x10,0x00,0x10,0x00,0x00,0x00,
-0x00,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xc0,0x0f,0x30,0x10,0x08,0x20,0x84,0x43,0x44,0x44,0x22,0x44,
-0x22,0x44,0x22,0x26,0xc2,0x1d,0x04,0x00,0x18,0x18,0xe0,0x07,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x20,0x00,
-0x50,0x00,0x50,0x00,0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,0xfe,0x03,0x02,0x02,
-0x01,0x04,0x01,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xfc,0x00,0x04,0x01,0x04,0x02,0x04,0x02,0x04,0x01,0xfc,0x00,
-0x04,0x01,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x01,0xfc,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,
-0x04,0x02,0x02,0x04,0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x04,0x04,0x02,
-0x0c,0x03,0xf0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xfc,0x00,0x04,0x03,0x04,0x02,0x04,0x04,0x04,0x04,0x04,0x04,
-0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x02,0x04,0x03,0xfc,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xfc,0x03,0x04,0x00,
-0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x01,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,
-0x04,0x00,0xfc,0x03,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xfc,0x03,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x01,
-0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,
-0x04,0x02,0x02,0x04,0x02,0x00,0x02,0x00,0xc2,0x07,0x02,0x04,0x02,0x04,0x04,0x06,
-0x0c,0x07,0xf0,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0xfc,0x03,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,
-0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x20,0x20,0x20,
-0x20,0x20,0x20,0x20,0x21,0x21,0x21,0x1e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x04,0x01,0x84,0x00,0x44,0x00,0x24,0x00,0x14,0x00,0x0c,0x00,
-0x14,0x00,0x24,0x00,0x44,0x00,0x84,0x00,0x04,0x01,0x04,0x02,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x00,0x04,0x00,
-0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,
-0x04,0x00,0xfc,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x04,0x04,0x04,0x04,0x0c,0x06,0x0c,0x06,0x14,0x05,0x14,0x05,
-0xa4,0x04,0xa4,0x04,0x44,0x04,0x44,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x01,0x0c,0x01,
-0x0c,0x01,0x14,0x01,0x14,0x01,0x24,0x01,0x24,0x01,0x44,0x01,0x44,0x01,0x84,0x01,
-0x84,0x01,0x04,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,0x02,0x04,0x02,0x04,
-0x02,0x04,0x02,0x04,0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xfc,0x00,0x04,0x01,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x01,0xfc,0x00,0x04,0x00,0x04,0x00,0x04,0x00,
-0x04,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,0x02,0x04,0x02,0x04,
-0x02,0x04,0x02,0x04,0x42,0x04,0x84,0x02,0x0c,0x03,0xf0,0x02,0x00,0x04,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xfc,0x00,0x04,0x01,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x01,0xfc,0x00,0x04,0x01,0x04,0x02,0x04,0x02,
-0x04,0x02,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xf8,0x00,0x04,0x01,0x02,0x02,0x02,0x00,0x04,0x00,0xf8,0x00,
-0x00,0x01,0x00,0x02,0x00,0x02,0x02,0x02,0x04,0x01,0xf8,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xff,0x01,0x10,0x00,
-0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,
-0x10,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x08,0x01,0xf0,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x04,0x01,0x04,
-0x02,0x02,0x02,0x02,0x04,0x01,0x04,0x01,0x88,0x00,0x88,0x00,0x50,0x00,0x50,0x00,
-0x20,0x00,0x20,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x01,0x40,0x81,0x40,0x82,0x20,0x82,0x20,0x42,0x21,0x44,0x11,
-0x44,0x11,0x28,0x0a,0x28,0x0a,0x10,0x04,0x10,0x04,0x10,0x04,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x04,0x02,0x02,
-0x04,0x01,0x88,0x00,0x50,0x00,0x20,0x00,0x20,0x00,0x50,0x00,0x88,0x00,0x04,0x01,
-0x02,0x02,0x01,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x01,0x04,0x02,0x02,0x04,0x01,0x88,0x00,0x50,0x00,0x20,0x00,
-0x20,0x00,0x20,0x00,0x20,0x00,0x20,0x00,0x20,0x00,0x20,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xfe,0x03,0x00,0x02,
-0x00,0x01,0x80,0x00,0x40,0x00,0x20,0x00,0x10,0x00,0x08,0x00,0x04,0x00,0x02,0x00,
-0x01,0x00,0xff,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x06,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x06,0x00,
-0x00,0x00,0x00,0x00,0x01,0x01,0x01,0x01,0x02,0x02,0x02,0x02,0x04,0x04,0x04,0x04,
-0x08,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x04,0x04,0x04,0x04,0x04,0x04,0x04,
-0x04,0x04,0x04,0x04,0x04,0x04,0x06,0x00,0x00,0x00,0x00,0x08,0x14,0x22,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xff,0x01,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x04,0x08,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x7c,0x00,0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x3a,0x00,
-0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x46,0x00,0x3a,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,
-0x44,0x82,0x02,0x02,0x02,0x82,0x44,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x80,0x00,0x80,0x00,0x80,0x00,0xb8,0x00,0xc4,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0xc4,0x00,0xb8,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0xfe,0x00,0x02,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x18,0x04,0x04,0x1f,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xb8,0x00,
-0xc4,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0xc4,0x00,0xb8,0x00,
-0x80,0x00,0x80,0x00,0x40,0x00,0x3c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x3a,0x00,0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x02,0x02,0x02,0x02,
-0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x01,0x00,0x00,0x00,0x00,0x02,0x02,0x02,0x42,
-0x22,0x12,0x0a,0x0e,0x12,0x22,0x42,0x82,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1a,0x03,
-0xa6,0x04,0x42,0x08,0x42,0x08,0x42,0x08,0x42,0x08,0x42,0x08,0x42,0x08,0x42,0x08,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3a,0x00,0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,
-0x44,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x3a,0x00,0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x46,0x00,0x3a,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xb8,0x00,
-0xc4,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0xc4,0x00,0xb8,0x00,
-0x80,0x00,0x80,0x00,0x80,0x00,0x80,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1a,
-0x06,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x3c,0x42,0x02,0x02,0x3c,0x40,0x40,0x42,0x3c,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x04,0x04,0x04,0x1f,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x08,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0xc4,0x00,0xb8,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x08,0x08,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x04,
-0x21,0x04,0x22,0x02,0x22,0x02,0x54,0x01,0x54,0x01,0x88,0x00,0x88,0x00,0x88,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x41,
-0x22,0x22,0x14,0x08,0x14,0x22,0x22,0x41,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x08,0x04,0x04,0x02,0x02,0x01,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7e,0x40,0x20,0x10,0x08,0x08,0x04,0x02,0x7e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x04,0x04,0x04,0x04,0x04,0x04,0x02,
-0x04,0x04,0x04,0x04,0x04,0x04,0x08,0x00,0x00,0x00,0x00,0x00,0x04,0x04,0x04,0x04,
-0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x02,0x04,0x04,0x04,0x04,0x04,0x04,0x08,0x04,0x04,0x04,0x04,0x04,0x04,0x02,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x8c,0x00,0x92,0x00,0x62,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x84,0x00,0x82,0x00,0x1f,0x00,0x02,0x00,
-0x0f,0x00,0x02,0x00,0x82,0x00,0x84,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x02,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0e,0x0e,0x0e,0x0e,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x0e,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x02,
-0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x80,0x00,0xb8,0x00,0x44,0x00,0xa2,0x00,0x22,0x00,0x12,0x00,
-0x0a,0x00,0x8a,0x00,0x44,0x00,0x3a,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x44,0x00,0x04,0x00,
-0x08,0x00,0x7e,0x00,0x08,0x00,0x08,0x00,0x08,0x00,0x04,0x00,0x04,0x00,0xfe,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x82,0x00,0x7c,0x00,0x44,0x00,0x44,0x00,0x44,0x00,0x7c,0x00,0x82,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x01,0x01,0x01,0x82,0x00,0x44,0x00,
-0xff,0x01,0x10,0x00,0x10,0x00,0xff,0x01,0x10,0x00,0x10,0x00,0x10,0x00,0x10,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x04,0x04,0x04,
-0x04,0x04,0x00,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x04,0x00,0x0c,0x00,0x12,0x00,0x22,0x00,
-0x44,0x00,0x88,0x00,0x90,0x00,0x60,0x00,0x40,0x00,0x44,0x00,0x38,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0a,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0xf8,0x00,0x06,0x03,0x02,0x02,0x71,0x04,0x89,0x04,0x09,0x04,0x09,0x04,0x89,0x04,
-0x71,0x04,0x02,0x06,0x06,0x03,0xf8,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x0c,0x10,0x1c,0x12,0x1c,0x00,0x1e,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x88,0x00,0x44,0x00,0x22,0x00,0x44,0x00,
-0x88,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0xfe,0x00,0x80,0x00,0x80,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x1f,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xf8,0x00,
-0x06,0x03,0x02,0x02,0x79,0x04,0x89,0x04,0x89,0x04,0x79,0x04,0x49,0x04,0x89,0x04,
-0x02,0x02,0x06,0x03,0xf8,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0xff,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0c,0x12,0x12,0x0c,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x10,0x00,0x10,0x00,
-0xfe,0x00,0x10,0x00,0x10,0x00,0x10,0x00,0x00,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0c,0x12,0x10,0x08,0x04,0x1e,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0c,0x12,0x08,0x10,
-0x12,0x0c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,
-0x04,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0x42,0xbe,
-0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0xfc,0x5e,0x5e,0x5e,0x5e,0x5c,0x50,0x50,
-0x50,0x50,0x50,0x50,0x50,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x08,0x08,0x06,
-0x00,0x00,0x00,0x00,0x04,0x06,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0c,0x12,0x12,0x12,0x0c,0x00,0x1e,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x22,0x00,0x44,0x00,
-0x88,0x00,0x44,0x00,0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x08,0x06,0x04,0x04,0x02,0x04,0x01,
-0x84,0x00,0x44,0x00,0x20,0x04,0x10,0x06,0x08,0x05,0x84,0x04,0x82,0x07,0x01,0x04,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x04,0x08,0x06,0x04,0x04,0x02,0x04,0x01,0x84,0x00,0x44,0x00,0x20,0x03,0x90,0x04,
-0x08,0x04,0x04,0x02,0x02,0x01,0x81,0x07,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0c,0x08,0x12,0x04,0x08,0x02,0x10,0x01,
-0x92,0x00,0x4c,0x00,0x20,0x04,0x10,0x06,0x08,0x05,0x84,0x04,0x82,0x07,0x01,0x04,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x10,0x00,
-0x08,0x00,0x04,0x00,0x02,0x00,0x02,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,
-0x08,0x00,0x10,0x00,0x20,0x00,0x00,0x00,0x20,0x00,0x20,0x00,0x50,0x00,0x50,0x00,
-0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,0xfe,0x03,0x02,0x02,0x01,0x04,0x01,0x04,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x40,0x00,0x20,0x00,0x00,0x00,
-0x20,0x00,0x20,0x00,0x50,0x00,0x50,0x00,0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,
-0xfe,0x03,0x02,0x02,0x01,0x04,0x01,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x20,0x00,0x50,0x00,0x88,0x00,0x00,0x00,0x20,0x00,0x20,0x00,0x50,0x00,0x50,0x00,
-0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,0xfe,0x03,0x02,0x02,0x01,0x04,0x01,0x04,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x18,0x01,0x24,0x01,0xc4,0x00,0x00,0x00,
-0x20,0x00,0x20,0x00,0x50,0x00,0x50,0x00,0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,
-0xfe,0x03,0x02,0x02,0x01,0x04,0x01,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x88,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x20,0x00,0x50,0x00,0x50,0x00,
-0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,0xfe,0x03,0x02,0x02,0x01,0x04,0x01,0x04,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x70,0x00,0x88,0x00,0x88,0x00,0x70,0x00,
-0x20,0x00,0x20,0x00,0x50,0x00,0x50,0x00,0x88,0x00,0x88,0x00,0x04,0x01,0x04,0x01,
-0xfe,0x03,0x02,0x02,0x01,0x04,0x01,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xe0,0x7f,0xa0,0x00,0x90,0x00,0x90,0x00,
-0x88,0x00,0x88,0x3f,0x84,0x00,0x84,0x00,0xfe,0x00,0x82,0x00,0x81,0x00,0x81,0x7f,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,
-0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,0x40,0x00,0x80,0x00,0x80,0x00,0x60,0x00,
-0x10,0x00,0x20,0x00,0x40,0x00,0x00,0x00,0xfc,0x03,0x04,0x00,0x04,0x00,0x04,0x00,
-0x04,0x00,0xfc,0x01,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x03,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x40,0x00,0x20,0x00,0x00,0x00,
-0xfc,0x03,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x01,0x04,0x00,0x04,0x00,
-0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x03,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x60,0x00,0x90,0x00,0x08,0x01,0x00,0x00,0xfc,0x03,0x04,0x00,0x04,0x00,0x04,0x00,
-0x04,0x00,0xfc,0x01,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x03,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x01,0x00,0x00,0x00,0x00,
-0xfc,0x03,0x04,0x00,0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x01,0x04,0x00,0x04,0x00,
-0x04,0x00,0x04,0x00,0x04,0x00,0xfc,0x03,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x01,0x02,0x04,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,
-0x00,0x00,0x00,0x00,0x10,0x08,0x04,0x00,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,
-0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x04,0x0a,0x11,0x00,0x04,0x04,0x04,0x04,
-0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,0x00,0x11,0x00,0x00,
-0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xfc,0x00,0x04,0x03,0x04,0x02,0x04,0x04,
-0x04,0x04,0x1f,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x04,0x02,0x04,0x03,0xfc,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x18,0x01,0x24,0x01,0xc4,0x00,0x00,0x00,
-0x04,0x01,0x0c,0x01,0x0c,0x01,0x14,0x01,0x14,0x01,0x24,0x01,0x24,0x01,0x44,0x01,
-0x44,0x01,0x84,0x01,0x84,0x01,0x04,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x10,0x00,0x20,0x00,0x40,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,
-0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x40,0x00,0x20,0x00,0x00,0x00,
-0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,
-0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x60,0x00,0x90,0x00,0x08,0x01,0x00,0x00,0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,
-0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x02,0x64,0x02,0xc4,0x01,0x00,0x00,
-0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,
-0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x08,0x01,0x00,0x00,0x00,0x00,0xf0,0x00,0x0c,0x03,0x04,0x02,0x02,0x04,
-0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x04,0x02,0x0c,0x03,0xf0,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x82,0x00,0x44,0x00,0x28,0x00,0x10,0x00,0x28,0x00,
-0x44,0x00,0x82,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xf0,0x04,0x0c,0x03,0x04,0x03,0x82,0x04,
-0x82,0x04,0x42,0x04,0x22,0x04,0x12,0x04,0x12,0x04,0x0c,0x02,0x0c,0x03,0xf2,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x20,0x00,0x40,0x00,0x00,0x00,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,
-0x04,0x02,0x04,0x02,0x08,0x01,0xf0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x80,0x00,0x40,0x00,0x20,0x00,0x00,0x00,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x08,0x01,0xf0,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x60,0x00,0x90,0x00,0x08,0x01,0x00,0x00,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,
-0x04,0x02,0x04,0x02,0x08,0x01,0xf0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x08,0x01,0x00,0x00,0x00,0x00,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,
-0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x04,0x02,0x08,0x01,0xf0,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x40,0x00,0x20,0x00,0x00,0x00,
-0x01,0x04,0x02,0x02,0x04,0x01,0x88,0x00,0x50,0x00,0x20,0x00,0x20,0x00,0x20,0x00,
-0x20,0x00,0x20,0x00,0x20,0x00,0x20,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x00,0x04,0x00,0x7c,0x00,0x84,0x00,
-0x04,0x01,0x04,0x01,0x04,0x01,0x04,0x01,0x84,0x00,0x7c,0x00,0x04,0x00,0x04,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x38,0x00,0x44,0x00,0x84,0x00,0x84,0x00,0x84,0x00,0x64,0x00,0x84,0x00,0x04,0x01,
-0x04,0x01,0x04,0x01,0x84,0x00,0x74,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,0x10,0x00,0x20,0x00,0x00,0x00,0x7c,0x00,
-0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,
-0x10,0x00,0x08,0x00,0x00,0x00,0x7c,0x00,0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x28,0x00,0x44,0x00,0x00,0x00,0x7c,0x00,
-0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x8c,0x00,
-0x92,0x00,0x62,0x00,0x00,0x00,0x7c,0x00,0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x44,0x00,0x00,0x00,0x00,0x00,0x7c,0x00,
-0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x7c,0x00,0x82,0x00,0x80,0x00,0x80,0x00,0xfc,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x7c,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7c,0x0e,
-0x82,0x11,0x80,0x20,0x80,0x20,0xfc,0x3f,0x82,0x00,0x82,0x20,0x82,0x11,0x7c,0x0e,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x38,
-0x44,0x82,0x02,0x02,0x02,0x82,0x44,0x38,0x10,0x20,0x20,0x18,0x00,0x00,0x00,0x00,
-0x00,0x00,0x08,0x00,0x10,0x00,0x20,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,
-0x82,0x00,0xfe,0x00,0x02,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x10,0x00,0x08,0x00,
-0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0xfe,0x00,0x02,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x10,0x00,0x28,0x00,0x44,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,
-0x82,0x00,0xfe,0x00,0x02,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x44,0x00,0x00,0x00,
-0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0xfe,0x00,0x02,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,
-0x02,0x04,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x04,0x02,0x01,0x00,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x05,0x00,0x00,0x02,0x02,0x02,0x02,0x02,
-0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x05,0x00,0x00,0x02,
-0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x16,0x00,0x08,0x00,0x14,0x00,0x20,0x00,0x38,0x00,0x44,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x8c,0x00,0x92,0x00,0x62,0x00,
-0x00,0x00,0x3a,0x00,0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x08,0x00,0x10,0x00,0x20,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x10,0x00,0x08,0x00,
-0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x10,0x00,0x28,0x00,0x44,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x8c,0x00,0x92,0x00,0x62,0x00,
-0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x44,0x00,0x00,0x00,0x00,0x00,0x38,0x00,0x44,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x44,0x00,0x38,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0xfe,0x00,0x00,0x00,0x00,0x00,0x10,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xb8,0x00,0x44,0x00,0xa2,0x00,
-0xa2,0x00,0x92,0x00,0x8a,0x00,0x8a,0x00,0x44,0x00,0x3a,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,0x10,0x00,0x20,0x00,
-0x00,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0xc4,0x00,0xb8,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x20,0x00,0x10,0x00,0x08,0x00,0x00,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0xc4,0x00,0xb8,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x28,0x00,0x44,0x00,
-0x00,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0xc4,0x00,0xb8,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x44,0x00,0x00,0x00,0x00,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x82,0x00,0x82,0x00,0xc4,0x00,0xb8,0x00,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x08,0x04,0x00,0x41,0x41,0x22,0x22,0x14,
-0x14,0x08,0x08,0x04,0x04,0x02,0x02,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-0x02,0x00,0x02,0x00,0x02,0x00,0x3a,0x00,0x46,0x00,0x82,0x00,0x82,0x00,0x82,0x00,
-0x82,0x00,0x82,0x00,0x46,0x00,0x3a,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x02,0x00,
-0x00,0x00,0x00,0x00,0x22,0x00,0x00,0x41,0x41,0x22,0x22,0x14,0x14,0x08,0x08,0x04,
-0x04,0x02,0x02,0x01 };
-const font_def_t sserif20 = { 256, 20, sserif20_widths, sserif20_offsets, sserif20_data };
+#endif
